AG-12088 Refactor out annotation types config
