+++ packages/ag-charts-community/src/util/vector.ts
@@ -14,7 +14,7 @@ export const Vec2 = {
     sub,
 };
 
-interface Vec2 {
+export interface Vec2 {
     x: number;
     y: number;
 }
